Convert from service to client
