+++ pom.xml
@@ -11,13 +11,11 @@
 
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>vertx-rabbitmq-service</artifactId>
+  <artifactId>vertx-rabbitmq-client</artifactId>
   <version>3.0.0-SNAPSHOT</version>
 
   <properties>
     <stack.version>3.0.0-SNAPSHOT</stack.version>
-    <service.proxy.version>3.0.0-SNAPSHOT</service.proxy.version>
-    <service.factory.version>3.0.0-SNAPSHOT</service.factory.version>
     <rabbitmq.client.version>3.4.3</rabbitmq.client.version>
   </properties>
 
@@ -37,14 +35,6 @@
     <dependency>
       <groupId>io.vertx</groupId>
       <artifactId>vertx-core</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.vertx</groupId>
-      <artifactId>vertx-service-proxy</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.vertx</groupId>
-      <artifactId>vertx-service-factory</artifactId>
     </dependency>
     <dependency>
       <groupId>io.vertx</groupId>
