--- src/main/resources/io.vertx.rabbitmq-service.json
@@ -1,11 +1,0 @@
-{
-  "main": "io.vertx.rabbitmq.RabbitMQServiceVerticle",
-  "options": {
-    "config": {
-      "address": "vertx.rabbitmq",
-      "connectionRetries": 5,
-      "connectionRetryDelay": 10000,
-      "connectionTimeout": 5000
-    }
-  }
-}
